e-com: fetch data and display on the screen.
