+++ src/components/pages/ProductsPage.js
@@ -2,6 +2,18 @@ import { useState, useEffect } from "react"
 import { Card } from "../organisms/Card";
 
 export const ProductsPage = () => {
-    return <main><Card />
+    const [product, setProduct] = useState([]);
+
+    useEffect(() => {
+        //api
+        fetch('https://fakestoreapi.com/products')
+            .then(data => data.json())
+            .then(result => setProduct(result))
+    }, [])
+    console.log({ product })
+    return <main>
+        {product.map((item) => {
+            return <Card itemDetail={item} />
+        })}
     </main>
 }
